+++ CHANGES.md
@@ -1,4 +1,25 @@
 # Release notes
+
+UP Fast Downward 0.2.0
+- FastDownwardPDDLPlanner:
+  - support anytime operation mode
+  - support specification of aliases and search configurations to be used by
+    Fast Downward
+  - support specification of translator options to be used by Fast Downward
+  - support specification of search time limit for Fast Downward (necessary
+    for usage of portfolios)
+  - support setting the log level used by Fast Downward
+- FastDownwardOptimalPDDLPlanner:
+  - support setting the log level used by Fast Downward
+- FastDownwardGrounder:
+  - follow changes made in UP 
+- credits: mention people behind the heuristics we use in the default
+  configurations
+- use latest Fast Downward version
+- common base class for satisficing and optimal solver engines
+- code quality improvements: typing information and style fixes
+- added jupyter notebook with examples
+
 UP Fast Downward 0.1.2
 - follow change in up:
   rename problem.env into problem.environment